LayoutTests/inspector-protocol/nmi-webaudio.html

<html>
<head>
<script type="text/javascript" src="../http/tests/inspector-protocol/resources/protocol-test.js"></script>
<script>

var sampleRate = 44100.0;
var renderLengthSeconds = 0.125;

function createSinWaveBuffer(context, lengthInSeconds, frequency)
{
    var audioBuffer = context.createBuffer(1, lengthInSeconds * sampleRate, sampleRate);
    var data = audioBuffer.getChannelData(0);
    for (var i = 0; i < audioBuffer.length; ++i)
        data[i] = Math.sin(frequency * 2 * Math.PI * i / sampleRate);
    return audioBuffer;
}

function runPageTest()
{
    testRunner.waitUntilDone();

    // Create offline audio context.
    var context = new webkitAudioContext(1, sampleRate * renderLengthSeconds, sampleRate);

    var bufferSource = context.createBufferSource();
    bufferSource.buffer = createSinWaveBuffer(context, renderLengthSeconds, 880);
    bufferSource.connect(context.destination);
    context.oncomplete = runTest;
    context.startRendering();
}

function test()
{
    function sizeForType(entry, typeName)
    {
        if (entry.name === typeName)
            return entry.size;

        if (!entry.children)
            return -1;

        for (var i = 0; i < entry.children.length; ++i) {
            var size = sizeForType(entry.children[i], typeName);
            if (size > -1)
                return size;
        }

        return -1;
    }

    function receivedSizeMoreThan(messageObject, typeName, expectedSize)
    {
        var entry = messageObject["result"]["distribution"];
        var size = sizeForType(entry, typeName);
        if (size > expectedSize)
            InspectorTest.log("PASS: received size for type " + typeName + " is more than " + expectedSize);
        else
            InspectorTest.log("FAIL: received size " + size + " for type " + typeName + " is less than " + expectedSize);
    }

    function handleMemorySnapshot(messageObject)
    {
        if ("error" in messageObject)
            InspectorTest.log(messageObject.error.message);
        else {
            receivedSizeMoreThan(messageObject, "Page.Audio", 15000000);
            receivedSizeMoreThan(messageObject, "Page.Audio.SharedData", 15000000);
        }

        InspectorTest.completeTest();
    }

    InspectorTest.sendCommand("Memory.getProcessMemoryDistribution", {}, handleMemorySnapshot);
}
</script>
</head>
<body onLoad="runPageTest();">
</body>
</html>
