LayoutTests/inspector-protocol/nmi-webaudio-expected.txt

PASS: received size for type Page.Audio is more than 15000000
PASS: received size for type Page.Audio.SharedData is more than 15000000
